% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{avail}
\alias{avail}
\title{What objects of this class are available}
\usage{
avail(classy, package = c("SWOMSE", "all"))
}
\arguments{
\item{classy}{A class of object (character string, e.g. 'Fleet')}

\item{package}{Character. Either 'SWOMSE' to return objects from
package `SWOMSE` and global environment (default) or 'all' to also returns
objects from the `MSEtool` and `DLMtool` packages.}
}
\value{
A character string of available objects of the specified class
}
\description{
Finds objects of the specified class in the global environment and `SWOMSE` package,
or from `SWOMSE`, `DLMtool`, and `MSEtool` packages.
}
\examples{
avail("OM")
}
