% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{AverageC}
\alias{AverageC}
\title{A simple constant catch MP using average historical catch}
\usage{
AverageC(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{reps}{The number of replicates to include}
}
\value{
An object of class `Rec` with management recommendations
}
\description{
Future TACs are set to the most average historical catch
}
\examples{
AverageC(1, SWOData)
}
