% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{basecase_catch}
\alias{basecase_catch}
\title{List of observed and predicted catch for base case OM}
\format{
An object of class \code{data.frame} with 748 rows and 7 columns.
}
\usage{
basecase_catch
}
\description{

}
\keyword{datasets}
