% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{basecase_CPUE}
\alias{basecase_CPUE}
\title{CPUE Fitting Information for base case OM}
\format{
An object of class \code{data.frame} with 336 rows and 7 columns.
}
\usage{
basecase_CPUE
}
\description{
Fitted CPUE indices for base case OM
}
\keyword{datasets}
