% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{basecase_DF}
\alias{basecase_DF}
\title{Overview of Base Case OM Parameters}
\format{
An object of class \code{data.frame} with 2 rows and 10 columns.
}
\usage{
basecase_DF
}
\description{
A dataframe of OM Parameters
}
\keyword{datasets}
