% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{basecase_LenDat}
\alias{basecase_LenDat}
\title{List of observed and predicted length data for base case OM}
\format{
An object of class \code{data.frame} with 8333 rows and 11 columns.
}
\usage{
basecase_LenDat
}
\description{

}
\keyword{datasets}
