% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{basecase_LH}
\alias{basecase_LH}
\title{Base Case OM Likelihoods}
\format{
An object of class \code{list} of length 2.
}
\usage{
basecase_LH
}
\description{
A list of likelihood values for base case OM
}
\keyword{datasets}
