% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{basecase_RefPoint}
\alias{basecase_RefPoint}
\title{Dataframe of Reference Points for base case OM}
\format{
An object of class \code{data.frame} with 1 rows and 8 columns.
}
\usage{
basecase_RefPoint
}
\description{
None converged OMs have Depletion=NA
}
\keyword{datasets}
