% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{basecase_Select}
\alias{basecase_Select}
\title{List of estimated selectivity-at-size for base case OM}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 693 rows and 9 columns.
}
\usage{
basecase_Select
}
\description{

}
\keyword{datasets}
