% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{basecase_TSbio}
\alias{basecase_TSbio}
\title{Time-Series Data of Predicted Spawning Biomass and Recruitment}
\format{
An object of class \code{data.frame} with 68 rows and 15 columns.
}
\usage{
basecase_TSbio
}
\description{
A list of predicted spawning biomass and recruitment for base case OM
}
\keyword{datasets}
