% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE.R
\name{CE}
\alias{CE}
\alias{CE_a}
\alias{CE_e}
\alias{CE_h}
\title{Constant Exploitation}
\usage{
CE(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  yrs = c(5, 3),
  ...
)

CE_a(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1.00732609275053,
  mc = 0.25,
  yrs = c(5, 3),
  ...
)

CE_e(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1.05251409774436,
  mc = 0.25,
  yrs = c(5, 3),
  ...
)

CE_h(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1.05535714285714,
  mc = 0.25,
  yrs = c(5, 3),
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{yrs}{numeric values length 2. The year index to calculate the historical mean catches and index,
and the years to smooth the estimate of current exploitation rate}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
This MP aims to keep the exploitation rate at a constant level. The current relative
exploitation rate is calculated as the mean of the catches from 2016:2020 divided
by the mean Combined Index over this same period.
}
\section{Functions}{
\itemize{
\item \code{CE_a()}: Tuned to PGK_6_10 = 0.6 across Reference OMs.

\item \code{CE_e()}: Tuned to PGK_6_10 = 0.51 across Reference OMs.

\item \code{CE_h()}: Tuned to LRP = 0.05 across Reference OMs.

}}
