% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{combine_MMSE}
\alias{combine_MMSE}
\title{Combine multiple MMSE objects together}
\usage{
combine_MMSE(MMSElist, name)
}
\arguments{
\item{MMSElist}{A list of objects of class `MMSE`}

\item{name}{The name for the `MMSE` object that is returned}
}
\value{
An object of class `MMSE`
}
\description{
Combine multiple MMSE objects together
}
