% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{ConstC}
\alias{ConstC}
\title{A simple constant catch MP}
\usage{
ConstC(x, Data, reps = 100)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{reps}{The number of replicates to include}
}
\value{
An object of class `Rec` with management recommendations
}
\description{
Future TACs are set to the most recent TAC
}
\examples{
ConstC(1, SWOData)
}
