% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{Data_DF}
\alias{Data_DF}
\title{Data Information}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22391 rows and 13 columns.
}
\usage{
Data_DF
}
\description{
A dataframe of fishery data
}
\keyword{datasets}
