% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{delete_tuning}
\alias{delete_tuning}
\title{Delete tuning and scoping objects for an MP}
\usage{
delete_tuning(MP_name, Tune_dir = "Tuning_Objects")
}
\arguments{
\item{MP_name}{Name of an `MP` function}

\item{Tune_dir}{Directory where scoping results are saved}
}
\value{
nothing
}
\description{
Delete tuning and scoping objects for an MP
}
