% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{document_MP}
\alias{document_MP}
\title{Document and Export Tuned CMPs}
\usage{
document_MP(MP, tune.dir)
}
\arguments{
\item{MP}{The name of the tuned CMP. The CMP function must be in an R script in 'R' directory with the file name
`MP.R`}

\item{tune.dir}{Directory where the tuning parameter objects (created with `TuneCMP`) are saved.}
}
\value{
Writes the tuned CMPs to `MP.R`
}
\description{
Document and Export Tuned CMPs
}
