% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_AH.R
\name{EA1}
\alias{EA1}
\alias{EA1_a}
\title{EA 1}
\usage{
EA1(
  x,
  Data,
  Index_ID = c(1, 5, 7),
  Index_CV = c(0.275, 0.219, 0.307),
  Data_Lag = 2,
  Interval = 3,
  yrsmth = 3,
  mc = 0.5,
  tunepar = 1,
  ...
)

EA1_a(
  x,
  Data,
  Index_ID = c(1, 5, 7),
  Index_CV = c(0.275, 0.219, 0.307),
  Data_Lag = 2,
  Interval = 3,
  yrsmth = 3,
  mc = 0.5,
  tunepar = 1.2,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{tunepar}{Parameter used for tuning}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
EA 1
}
\section{Functions}{
\itemize{
\item \code{EA1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

}}
