% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_Diagnostics.R
\name{get_CI_Index}
\alias{get_CI_Index}
\title{Get the Combined Index for an MMSE object}
\usage{
get_CI_Index(MMSE)
}
\arguments{
\item{MMSE}{Object of class `MMSE`}
}
\value{
a data.frame
}
\description{
Get the Combined Index for an MMSE object
}
