% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_Diagnostics.R
\name{get_Index}
\alias{get_Index}
\alias{AddIndList}
\title{Get the Index from `PPD@Ind` slot in an MMSE object}
\usage{
get_Index(MMSE, Name = "Combined Index", Stock = 1, Fleet = 1)

AddIndList(MOM)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{Name}{Name for the index.}

\item{Stock}{Stock index number. Default to 1}

\item{Fleet}{Fleet index number. Default to 1}
}
\value{
A data.frame
}
\description{
Get the Index from `PPD@Ind` slot in an MMSE object
}
\section{Functions}{
\itemize{
\item \code{AddIndList()}: Get the additional indices

}}
