% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{get_tune_PM}
\alias{get_tune_PM}
\title{Get Performance Metrics from Tuning Objects}
\usage{
get_tune_PM(tune_dir = "Tuning_Objects", tune_PM = "PGK_short")
}
\description{
Get Performance Metrics from Tuning Objects
}
