% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{ImportAll}
\alias{ImportAll}
\title{Loop over directories and import SS OMs}
\usage{
ImportAll(
  rootdir = "C:/Users/Adrian/Dropbox/SWO_MSE/OMs/SS/2018_GRID",
  Gen = 1:2,
  overwrite = TRUE
)
}
\arguments{
\item{rootdir}{root directory containing sub-directories with SS output files}

\item{Gen}{Gen Genders to import. Numeric. `Gen=1` for first gender (usually female),
`Gen=2` for second gender (usually male), etc. `Gen=1:2` (default) uses average from both genders}

\item{overwrite}{Logical. Overwrite existing files?}
}
\value{
nothing
}
\description{
Loop over directories and import SS OMs
}
\keyword{internal}
