% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{ImportSSOM}
\alias{ImportSSOM}
\title{Import a SS OM and save to package data directory}
\usage{
ImportSSOM(SSdir, Gen = 1:2, overwrite = TRUE)
}
\arguments{
\item{SSdir}{File path to SS output files}

\item{Gen}{Gen Genders to import. Numeric. `Gen=1` for first gender (usually female),
`Gen=2` for second gender (usually male), etc. `Gen=1:2` (default) uses average from both genders}

\item{overwrite}{Logical. Overwrite existing files?}
}
\value{
nothing
}
\description{
Import a SS OM and save to package data directory
}
\keyword{internal}
