% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_Diagnostics.R
\name{Index_Diagnostic}
\alias{Index_Diagnostic}
\title{Diagnostic Plots for the Indices}
\usage{
Index_Diagnostic(MMSE, Index = "CI")
}
\arguments{
\item{MMSE}{Object of class `MMSE`}

\item{Index}{'CI' for Combined Index or numeric value for one of the Additional Indices (not currently implemented)}
}
\value{
A list of ggplot objects
}
\description{
Diagnostic Plots for the Indices
}
