% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_Diagnostics.R
\name{Index_Diagnostic_CI}
\alias{Index_Diagnostic_CI}
\title{Diagnostic Plots for the Combined Index}
\usage{
Index_Diagnostic_CI(MMSE, fit.years = 1999:2020, nsims = 6)
}
\arguments{
\item{MMSE}{Object of class `MMSE`}

\item{fit.years}{Years where the index statistical properties are calculated}

\item{nsims}{Maximum number of simulations to plot}
}
\value{
A list of `ggplot` objects
}
\description{
Diagnostic Plots for the Combined Index
}
