% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IR1.R, R/IR2.R
\name{IR1}
\alias{IR1}
\alias{IR1_a}
\alias{IR1_b}
\alias{IR1_c}
\alias{IR2}
\title{Index Ratio 1}
\usage{
IR1(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  yrs = c(3, 3),
  ...
)

IR1_a(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 0.791,
  mc = 0.25,
  yrs = c(3, 3),
  ...
)

IR1_b(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 0.929,
  mc = 0.25,
  yrs = c(3, 3),
  ...
)

IR1_c(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 0.964,
  mc = 0.25,
  yrs = c(3, 3),
  ...
)

IR2(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  yrs = c(2, 2),
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{yrs}{numeric vector length 2. The years to calculate the mean index. Ratio is
calculated as mean(ind[yrs[1]])/mean(ind[yrs[2]])}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
This MP adjusts the TAC based on ratio of the mean index over the last 3 years (y-2):y
to the mean index over the 3 years before that (y-5):(y-3)
}
\section{Functions}{
\itemize{
\item \code{IR1_a()}: Tuned to PGK_med = 0.6 across Reference OMs.

\item \code{IR1_b()}: Tuned to PGK_long = 0.6 across Reference OMs.

\item \code{IR1_c()}: Tuned to PGK_30 = 0.6 across Reference OMs.

\item \code{IR2()}: Same as IR1, but mean index values are calculated over last 2 years and the previous 2 years before that

}}
