% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IT.R
\name{IT1}
\alias{IT1}
\title{Index Targeting 1}
\usage{
IT1(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  change = 0.1,
  smooth_yr = 3,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}

\item{yrs}{numeric vector length 2. The years to calculate the mean index. Ratio is
calculated as mean(ind[yrs[1]])/mean(ind[yrs[2]])}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
Index Targeting 1
}
