% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{Itarg1}
\alias{Itarg1}
\title{A simple index targeting MP}
\usage{
Itarg1(x, Data, reps = 100, yrsmth = 5, mc = 0.05)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{reps}{The number of replicates to include}

\item{yrsmth}{The number of recent years to average index over}

\item{mc}{The maximum fractional change in the TAC among years}
}
\value{
An object of class `Rec` with management recommendations
}
\description{
A simple index targeting MP
}
\examples{
 Itarg1(1, SWOData)
}
