% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{Itarg2}
\alias{Itarg2}
\title{A second simple index targeting MP}
\usage{
Itarg2(x, Data, reps = 100, Imulti = 1.5)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{reps}{The number of replicates to include}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is compared with recent levels}
}
\value{
An object of class `Rec` with management recommendations
}
\description{
A management procedure that incrementally adjusts the TAC (starting from
reference level that is a fraction of mean recent catches)
to reach a target CPUE / relative abundance index.
}
\details{
Based on the index target MP developed by Geromont and Butterworth (2014)
and evaluated in Carruthers et al. (2015).
}
\examples{
Itarg2(1, SWOData)

}
\references{
Carruthers et al. 2015. Performance evaluation of simple
management procedures. ICES J. Mar Sci. 73, 464-482.

Geromont, H.F., Butterworth, D.S. 2014. Generic management procedures for
data-poor fisheries; forecasting with few data. ICES J. Mar. Sci. 72, 251-261.
doi:10.1093/icesjms/fst232
}
