% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{ITarget_1}
\alias{ITarget_1}
\title{Targeting Model-Free Example MP}
\usage{
ITarget_1(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  yrsmth = 5,
  mc = 0.2,
  multi = 1,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{yrsmth}{The number of recent years to average index over}

\item{mc}{The maximum fractional change in the TAC among years}

\item{multi}{Multiplier for the index target from the average index}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec`
}
\description{
A example index-targeting MP that iteratively adjusts the TAC
based on the ratio of recent to overall index.
}
