% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{ITarget_2}
\alias{ITarget_2}
\title{Incremental Index Target MP}
\usage{
ITarget_2(
  x,
  Data,
  Data_Lag = 2,
  Interval = 3,
  yrsmth = 5,
  xx = 0,
  Imulti = 1.5,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{yrsmth}{The number of recent years to average index over}

\item{xx}{Parameter controlling the fraction of mean catch to start using in first year}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is compared with recent levels}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec`
}
\description{
A management procedure that incrementally adjusts the TAC
(starting from reference level that is a fraction of mean recent catches)
to reach a target CPUE / relative abundance index
See `?Itarget1` for more information.
}
