% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{LenDat_List}
\alias{LenDat_List}
\title{List of observed and predicted length data for each OM}
\format{
An object of class \code{list} of length 288.
}
\usage{
LenDat_List
}
\description{

}
\keyword{datasets}
