% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{Like_List}
\alias{Like_List}
\title{OM Likelihoods}
\format{
An object of class \code{list} of length 288.
}
\usage{
Like_List
}
\description{
A list of likelihood values for each OM
}
\keyword{datasets}
