% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{MakePerformanceDF}
\alias{MakePerformanceDF}
\title{Create Performance Dataframe}
\usage{
MakePerformanceDF(MSE)
}
\arguments{
\item{MSE}{An object of class `MMSE`}
}
\value{
A dataframe
}
\description{
Creates a dataframe with F/FMSY, SB/SBMSY, TAC, and Catch for each simulation, year, and MP
}
