% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Reports.R
\name{OM_Diagnostic_Report}
\alias{OM_Diagnostic_Report}
\title{Generate Diagnostic Report for specific OM}
\usage{
OM_Diagnostic_Report(OM, openFile = TRUE, quiet = FALSE, dev = FALSE)
}
\arguments{
\item{OM}{Name of the OM; either 'base_case' or 1 - 288}

\item{openFile}{Logical. Open the report in the browser?}

\item{quiet}{Logical. Suppress printing of the pandoc command line}

\item{dev}{Logical. Use the development version of RMD file?}
}
\value{
Nothing
}
\description{
Generate Diagnostic Report for specific OM
}
