% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_Reports.R
\name{OM_Summary_Report}
\alias{OM_Summary_Report}
\title{Generate OM Summary Report}
\usage{
OM_Summary_Report(openFile = TRUE, quiet = FALSE, dev = FALSE)
}
\arguments{
\item{openFile}{Logical. Open the report in the browser?}

\item{quiet}{Logical. Suppress printing of the pandoc command line}

\item{dev}{Logical. Use the development version of RMD file?}
}
\value{
Nothing
}
\description{
Creates a HTML report with tables and figures summarizing the key properties
of the OMs from the base case and uncertainty grid.
}
