% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{OMs_DF}
\alias{OMs_DF}
\title{Overview of OM Parameters}
\format{
An object of class \code{data.frame} with 288 rows and 10 columns.
}
\usage{
OMs_DF
}
\description{
A dataframe of OM Parameters
}
\keyword{datasets}
