% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Plot_Scope}
\alias{Plot_Scope}
\title{Plot the results of scoping for an MP}
\usage{
Plot_Scope(MP, Tune_dir = "Tuning_Objects")
}
\arguments{
\item{MP}{Name of the MP}

\item{Tune_dir}{Directory where scoping results are saved}
}
\value{
invisible list with ggplot object and data.frame
}
\description{
Plot the results of scoping for an MP
}
