% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Plot_Tune}
\alias{Plot_Tune}
\title{Plot the results of tuning for an MP}
\usage{
Plot_Tune(
  MP,
  PM = "PGK_6_10",
  Target = 0.6,
  Tune_dir = "Tuning_Objects",
  plot = TRUE
)
}
\arguments{
\item{MP}{Name of the MP}

\item{PM}{name of the performance metric}

\item{Target}{target for the PM}

\item{Tune_dir}{Directory where scoping results are saved}

\item{plot}{Logical. Show the plot?}
}
\value{
the tuning parameter
}
\description{
Plot the results of tuning for an MP
}
