% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{PM_table}
\alias{PM_table}
\title{Calculate Performance Metrics}
\usage{
PM_table(MMSE, PMs = NULL, msg = TRUE)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{PMs}{Optional. Names of `PM` functions to use. Otherwise all available PMs will be calculated.}

\item{msg}{Logical. Print messages?}
}
\value{
A `data.frame`
}
\description{
Calculate Performance Metrics
}
