% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{PMs}
\alias{PMs}
\alias{PGK_short}
\alias{PBMSY}
\alias{PBMSY_short}
\alias{PGK_med}
\alias{PGK_long}
\alias{PGK}
\alias{PGK_30}
\alias{POF}
\alias{PNOF}
\alias{LRP_short}
\alias{nLRP_short}
\alias{LRP_med}
\alias{nLRP_med}
\alias{LRP_long}
\alias{nLRP_long}
\alias{LRP}
\alias{nLRP}
\alias{TAC1}
\alias{AvTAC_short}
\alias{rAvTAC_short}
\alias{AvTAC_med}
\alias{rAvTAC_med}
\alias{AvTAC_long}
\alias{rAvTAC_long}
\alias{VarC}
\title{Swordfish Performance Metrics functions}
\usage{
PGK_short(MMSEobj = NULL, Ref = 1, Yrs = c(3, 12))

PBMSY(MMSEobj = NULL, Ref = 1, Yrs = c(3, 32))

PBMSY_short(MMSEobj = NULL, Ref = 1, Yrs = c(3, 12))

PGK_med(MMSEobj = NULL, Ref = 1, Yrs = c(13, 22))

PGK_long(MMSEobj = NULL, Ref = 1, Yrs = c(23, 32))

PGK(MMSEobj = NULL, Ref = 1, Yrs = c(3, 32))

PGK_30(MMSEobj = NULL, Ref = 1, Yrs = c(32, 32))

POF(MMSEobj = NULL, Ref = 1, Yrs = c(3, 32))

PNOF(MMSEobj = NULL, Ref = 1, Yrs = c(3, 32))

LRP_short(MMSEobj = NULL, Ref = 0.4, Yrs = c(3, 12))

nLRP_short(MMSEobj = NULL, Ref = 0.4, Yrs = c(3, 12))

LRP_med(MMSEobj = NULL, Ref = 0.4, Yrs = c(13, 22))

nLRP_med(MMSEobj = NULL, Ref = 0.4, Yrs = c(13, 22))

LRP_long(MMSEobj = NULL, Ref = 0.4, Yrs = c(23, 32))

nLRP_long(MMSEobj = NULL, Ref = 0.4, Yrs = c(23, 32))

LRP(MMSEobj = NULL, Ref = 0.4, Yrs = c(3, 32))

nLRP(MMSEobj = NULL, Ref = 0.4, Yrs = c(3, 32))

TAC1(MMSEobj = NULL, Ref = 1, Yrs = c(3, 3))

AvTAC_short(MMSEobj = NULL, Ref = NULL, Yrs = c(3, 12))

rAvTAC_short(MMSEobj = NULL, Ref = NULL, Yrs = c(3, 12))

AvTAC_med(MMSEobj = NULL, Ref = 1, Yrs = c(12, 22))

rAvTAC_med(MMSEobj = NULL, Ref = 1, Yrs = c(13, 22))

AvTAC_long(MMSEobj = NULL, Ref = 1, Yrs = c(23, 32))

rAvTAC_long(MMSEobj = NULL, Ref = 1, Yrs = c(23, 32))

VarC(MMSEobj = NULL, Ref = 1, Yrs = c(3, 32))
}
\arguments{
\item{MMSEobj}{An object of class `MMSE`}

\item{Ref}{Reference point used in the performance metrics (e.g., 0.5BMSY)}

\item{Yrs}{Years the performance metric is calculated over}
}
\value{
An object of class `PM`
}
\description{
Swordfish Performance Metrics functions
}
\section{Functions}{
\itemize{
\item \code{PGK_short()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Years 1-10 (2025-2034)

\item \code{PBMSY()}: Probability of SB>SBMSY over all years

\item \code{PBMSY_short()}: Probability of SB>SBMSY in Years 1-10 (2025-2034)

\item \code{PGK_med()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Years 11-20 (2035-2044)

\item \code{PGK_long()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Years 21-30 (2045-2054)

\item \code{PGK()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) over all years (2025-2054)

\item \code{PGK_30()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Year 30 (2054)

\item \code{POF()}: Probability of Overfishing (F>FMSY) over all years (2025-2054)

\item \code{PNOF()}: Probability of Not Overfishing (F<FMSY) over all years (2025-2054)

\item \code{LRP_short()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any of the first 10 years (2025-2034)

\item \code{nLRP_short()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any of the first 10 years (2025-2034)

\item \code{LRP_med()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any of years 11-20 (2035-2044)

\item \code{nLRP_med()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any of years 11-20 (2035-2044)

\item \code{LRP_long()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any of years 21-30 (2045-2054)

\item \code{nLRP_long()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any of years 21-30 (2045-2054)

\item \code{LRP()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any year (2025-2054)

\item \code{nLRP()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any year (2025-2054)

\item \code{TAC1()}: TAC in the First Implementation Year (2025)

\item \code{AvTAC_short()}: Median TAC (t) over years 1-10 (2025-2034)

\item \code{rAvTAC_short()}: Median TAC relative to MSY over years 1-10 (2025-2034)

\item \code{AvTAC_med()}: Median TAC (t) over years 11-20 (2035-2044)

\item \code{rAvTAC_med()}: Median TAC relative to MSY over years 11-20 (2035-2044)

\item \code{AvTAC_long()}: Median TAC (t) over years 21-30 (2045-2054)

\item \code{rAvTAC_long()}: Median TAC relative to MSY over years 21-30 (2045-2054)

\item \code{VarC()}: Mean variation in TAC (\%) between management cycles over all years and simulations

}}
\concept{Safety}
\concept{Stability}
\concept{Status}
\concept{Yield}
