% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{Proj_plot}
\alias{Proj_plot}
\title{Projection plots}
\usage{
Proj_plot(MMSE, sl = "SB_SBMSY", quants = c(0.05, 0.95))
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{sl}{Slot name. Either: `SB_SBMSY`, `F_FMSY`, `Catch`, or `Removals`}

\item{quants}{Quantiles to plot}
}
\value{
A `ggplot2` object
}
\description{
Projection plots
}
