% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{readSS}
\alias{readSS}
\title{Read SS output and data}
\usage{
readSS(SS.dir, i = "base_case")
}
\arguments{
\item{SS.dir}{Full file path to SS output files}
}
\value{
A named list
}
\description{
Read SS output and data
}
\keyword{internal}
