% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{RefPoint_DF}
\alias{RefPoint_DF}
\title{Dataframe of Reference Points for each OM}
\format{
An object of class \code{data.frame} with 288 rows and 8 columns.
}
\usage{
RefPoint_DF
}
\description{
None converged OMs have Depletion=NA
}
\keyword{datasets}
