% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{saveBaseCaseObjects}
\alias{saveBaseCaseObjects}
\title{Import base case SS objects and save to root}
\usage{
saveBaseCaseObjects(OMbase.dir)
}
\arguments{
\item{OMbase.dir}{}
}
\value{
Nothing
}
\description{
Import base case SS objects and save to root
}
\keyword{internal}
