% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{saveGridObjects}
\alias{saveGridObjects}
\title{Import SS objects from uncertainty grid and save to root}
\usage{
saveGridObjects(OM.dirs)
}
\arguments{
\item{OM.dirs}{}
}
\value{
Nothing
}
\description{
Import SS objects from uncertainty grid and save to root
}
\keyword{internal}
