% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{saveTempObject}
\alias{saveTempObject}
\title{Import SS objects and save to root}
\usage{
saveTempObject(OM.dirs)
}
\arguments{
\item{OM.dirs}{}
}
\value{
Nothing
}
\description{
Import SS objects and save to root
}
\keyword{internal}
