% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Scope}
\alias{Scope}
\title{Scope the Performance of a CMP across different tuning values}
\usage{
Scope(
  MP_name,
  Tuning_OMs,
  Hist_dir = "Hist_Objects/Reference",
  Tune_dir = "Tuning_Objects",
  test_vals = NULL,
  parallel = TRUE
)
}
\arguments{
\item{MP_name}{Name of an `MP` function}

\item{Tuning_OMs}{Names of the `MOM` objects to tune over}

\item{Hist_dir}{Directory where `.hist` objects are saved}

\item{Tune_dir}{Directory where scoping results will be saved}

\item{test_vals}{Optional test values to use}

\item{parallel}{Logical. Run in parallel? (recommended)}
}
\value{
An invisible data.frame
}
\description{
Calculates performance metrics for a range of tuning values. Uses the Reference
OMs with 20 simulations
}
