% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{Select_List}
\alias{Select_List}
\title{List of estimated selectivity-at-size for each OM}
\format{
An object of class \code{list} of length 288.
}
\usage{
Select_List
}
\description{

}
\keyword{datasets}
