% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny.R
\name{Shiny}
\alias{Shiny}
\title{Run the SWO MSE Shiny Application}
\usage{
Shiny(...)
}
\description{
\code{Shiny} runs the SWO MSE Shiny Application
}
\references{
Modified from Deal Attali's code: \url{http://deanattali.com/2015/04/21/r-package-shiny-app/}
}
