% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP25.R, R/SP25_Fox.R
\name{SP}
\alias{SP}
\alias{SP_Fox}
\title{A Surplus Production model with a Harvest Control Rule that linearly reduces F}
\usage{
SP(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = NA, ...)

SP_Fox(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = NA, ...)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
The SP model assumes a Schaefer production model
}
\section{Functions}{
\itemize{
\item \code{SP_Fox()}: Same as SP, but uses the Fox production model

}}
