% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP1.R, R/SP2.R, R/SP3.R
\name{SP1}
\alias{SP1}
\alias{SP2}
\alias{SP3}
\title{A Surplus Production model with a Harvest Control Rule that linearly reduces F}
\usage{
SP1(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)

SP2(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)

SP3(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
The SP model assumes a Schaefer production model
}
\section{Functions}{
\itemize{
\item \code{SP2()}: Same as SP1, but uses the Fox production model

\item \code{SP3()}: Fox production model with

}}
