% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{SP_1}
\alias{SP_1}
\title{ASPIC-like Surplus Production Assessment Model with TAC = MSY}
\usage{
SP_1(x, Data, Data_Lag = 2, Interval = 3, ...)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec`
}
\description{
An example model-based CMP that uses the `SAMtool::SP` function, which is similar
to ASPIC. It sets the TAC to the estimated MSY.
}
