% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{SPICT_1}
\alias{SPICT_1}
\title{SPiCT Assessment Model with TAC = MSY}
\usage{
SPICT_1(x, Data, Data_Lag = 2, Interval = 3, ...)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec`
}
\description{
An example model-based CMP that uses the `MSEextra::spict` function as a wrapper
to the `spict` package to assess the stock.
It sets the TAC = estimated MSY.
}
\details{
Note: this MP requires the `MSEextra` package to be installed. Run `MSEextra()` and
then `library(MSEextra)`
}
