% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{STC}
\alias{STC}
\title{Average Yield in the first 10 Years}
\usage{
STC(MSEobj = NULL, Ref = 1, Yrs = 10)
}
\arguments{
\item{MSEobj}{An object of class `MSE`}

\item{Ref}{Reference point for calculating the performance metric.}

\item{Yrs}{Numeric vector of length 2 with year indices to summarize performance.
If NULL, the performance is summarized over all projection years.}
}
\value{
An object of class `PMobj`
}
\description{
See `?LTY` for more information.
}
\examples{
\dontrun{
MSE <- runMSE()
STC(MSE)
}
}
