% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{Sub_MMSE}
\alias{Sub_MMSE}
\title{Subset an MMSE object by MP}
\usage{
Sub_MMSE(MMSE, MPs = NA)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{name}{The names of the MPs to keep}
}
\value{
An object of class `MMSE`
}
\description{
Subset an MMSE object by MP
}
