% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen.r
\docType{data}
\name{SWO-OMs}
\alias{SWO-OMs}
\alias{MOM_0}
\alias{MOM_1}
\alias{MOM_2}
\alias{MOM_3}
\alias{MOM_4}
\alias{MOM_5}
\alias{MOM_6}
\alias{MOM_7}
\alias{MOM_8}
\alias{MOM_9}
\alias{MOM_10}
\alias{MOM_11}
\alias{MOM_12}
\alias{MOM_13}
\alias{MOM_14}
\alias{MOM_15}
\alias{MOM_16}
\alias{MOM_17}
\alias{MOM_18}
\alias{MOM_19}
\alias{MOM_20}
\alias{MOM_21}
\alias{MOM_22}
\alias{MOM_23}
\alias{MOM_24}
\alias{MOM_25}
\alias{MOM_26}
\alias{MOM_27}
\alias{MOM_28}
\alias{MOM_29}
\alias{MOM_30}
\alias{MOM_31}
\alias{MOM_32}
\alias{MOM_33}
\alias{MOM_34}
\alias{MOM_35}
\alias{MOM_36}
\alias{MOM_37}
\alias{MOM_38}
\alias{MOM_39}
\alias{MOM_40}
\alias{MOM_41}
\alias{MOM_42}
\alias{MOM_43}
\alias{MOM_44}
\alias{MOM_45}
\alias{MOM_46}
\alias{MOM_47}
\alias{MOM_48}
\alias{MOM_49}
\alias{MOM_50}
\alias{MOM_51}
\alias{MOM_52}
\alias{MOM_53}
\alias{MOM_54}
\alias{MOM_55}
\alias{MOM_56}
\alias{MOM_57}
\alias{MOM_58}
\alias{MOM_59}
\alias{MOM_60}
\alias{MOM_61}
\alias{MOM_62}
\alias{MOM_63}
\alias{MOM_64}
\alias{MOM_65}
\alias{MOM_66}
\alias{MOM_67}
\alias{MOM_68}
\alias{MOM_69}
\alias{MOM_70}
\alias{MOM_71}
\alias{MOM_72}
\alias{MOM_73}
\alias{MOM_74}
\alias{MOM_75}
\alias{MOM_76}
\alias{MOM_77}
\alias{MOM_78}
\alias{MOM_79}
\alias{MOM_80}
\alias{MOM_81}
\alias{MOM_82}
\alias{MOM_83}
\alias{MOM_84}
\alias{MOM_85}
\alias{MOM_86}
\alias{MOM_87}
\alias{MOM_88}
\alias{MOM_89}
\alias{MOM_90}
\alias{MOM_91}
\alias{MOM_92}
\alias{MOM_93}
\alias{MOM_94}
\alias{MOM_95}
\alias{MOM_96}
\alias{MOM_97}
\alias{MOM_98}
\alias{MOM_99}
\alias{MOM_100}
\alias{MOM_101}
\alias{MOM_102}
\alias{MOM_103}
\alias{MOM_104}
\alias{MOM_105}
\alias{MOM_106}
\alias{MOM_107}
\alias{MOM_108}
\alias{MOM_109}
\alias{MOM_110}
\alias{MOM_111}
\alias{MOM_112}
\alias{MOM_113}
\alias{MOM_114}
\alias{MOM_115}
\alias{MOM_116}
\alias{MOM_117}
\alias{MOM_118}
\alias{MOM_119}
\alias{MOM_120}
\alias{MOM_121}
\alias{MOM_122}
\alias{MOM_123}
\alias{MOM_124}
\alias{MOM_125}
\alias{MOM_126}
\alias{MOM_127}
\alias{MOM_128}
\alias{MOM_129}
\alias{MOM_130}
\alias{MOM_131}
\alias{MOM_132}
\alias{MOM_133}
\alias{MOM_134}
\alias{MOM_135}
\alias{MOM_136}
\alias{MOM_137}
\alias{MOM_138}
\alias{MOM_139}
\alias{MOM_140}
\alias{MOM_141}
\alias{MOM_142}
\alias{MOM_143}
\alias{MOM_144}
\alias{MOM_145}
\alias{MOM_146}
\alias{MOM_147}
\alias{MOM_148}
\alias{MOM_149}
\alias{MOM_150}
\alias{MOM_151}
\alias{MOM_152}
\alias{MOM_153}
\alias{MOM_154}
\alias{MOM_155}
\alias{MOM_156}
\alias{MOM_157}
\alias{MOM_158}
\alias{MOM_159}
\alias{MOM_160}
\alias{MOM_161}
\alias{MOM_162}
\alias{MOM_163}
\alias{MOM_164}
\alias{MOM_165}
\alias{MOM_166}
\alias{MOM_167}
\alias{MOM_168}
\alias{MOM_169}
\alias{MOM_170}
\alias{MOM_171}
\alias{MOM_172}
\alias{MOM_173}
\alias{MOM_174}
\alias{MOM_175}
\alias{MOM_176}
\alias{MOM_177}
\alias{MOM_178}
\alias{MOM_179}
\alias{MOM_180}
\alias{MOM_181}
\alias{MOM_182}
\alias{MOM_183}
\alias{MOM_184}
\alias{MOM_185}
\alias{MOM_186}
\alias{MOM_187}
\alias{MOM_188}
\alias{MOM_189}
\alias{MOM_190}
\alias{MOM_191}
\alias{MOM_192}
\alias{MOM_193}
\alias{MOM_194}
\alias{MOM_195}
\alias{MOM_196}
\alias{MOM_197}
\alias{MOM_198}
\alias{MOM_199}
\alias{MOM_200}
\alias{MOM_201}
\alias{MOM_202}
\alias{MOM_203}
\alias{MOM_204}
\alias{MOM_205}
\alias{MOM_206}
\alias{MOM_207}
\alias{MOM_208}
\alias{MOM_209}
\alias{MOM_210}
\alias{MOM_211}
\alias{MOM_212}
\alias{MOM_213}
\alias{MOM_214}
\alias{MOM_215}
\alias{MOM_216}
\title{Operating models}
\format{
An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.
}
\usage{
MOM_0

MOM_1

MOM_2

MOM_3

MOM_4

MOM_5

MOM_6

MOM_7

MOM_8

MOM_9

MOM_10

MOM_11

MOM_12

MOM_13

MOM_14

MOM_15

MOM_16

MOM_17

MOM_18

MOM_19

MOM_20

MOM_21

MOM_22

MOM_23

MOM_24

MOM_25

MOM_26

MOM_27

MOM_28

MOM_29

MOM_30

MOM_31

MOM_32

MOM_33

MOM_34

MOM_35

MOM_36

MOM_37

MOM_38

MOM_39

MOM_40

MOM_41

MOM_42

MOM_43

MOM_44

MOM_45

MOM_46

MOM_47

MOM_48

MOM_49

MOM_50

MOM_51

MOM_52

MOM_53

MOM_54

MOM_55

MOM_56

MOM_57

MOM_58

MOM_59

MOM_60

MOM_61

MOM_62

MOM_63

MOM_64

MOM_65

MOM_66

MOM_67

MOM_68

MOM_69

MOM_70

MOM_71

MOM_72

MOM_73

MOM_74

MOM_75

MOM_76

MOM_77

MOM_78

MOM_79

MOM_80

MOM_81

MOM_82

MOM_83

MOM_84

MOM_85

MOM_86

MOM_87

MOM_88

MOM_89

MOM_90

MOM_91

MOM_92

MOM_93

MOM_94

MOM_95

MOM_96

MOM_97

MOM_98

MOM_99

MOM_100

MOM_101

MOM_102

MOM_103

MOM_104

MOM_105

MOM_106

MOM_107

MOM_108

MOM_109

MOM_110

MOM_111

MOM_112

MOM_113

MOM_114

MOM_115

MOM_116

MOM_117

MOM_118

MOM_119

MOM_120

MOM_121

MOM_122

MOM_123

MOM_124

MOM_125

MOM_126

MOM_127

MOM_128

MOM_129

MOM_130

MOM_131

MOM_132

MOM_133

MOM_134

MOM_135

MOM_136

MOM_137

MOM_138

MOM_139

MOM_140

MOM_141

MOM_142

MOM_143

MOM_144

MOM_145

MOM_146

MOM_147

MOM_148

MOM_149

MOM_150

MOM_151

MOM_152

MOM_153

MOM_154

MOM_155

MOM_156

MOM_157

MOM_158

MOM_159

MOM_160

MOM_161

MOM_162

MOM_163

MOM_164

MOM_165

MOM_166

MOM_167

MOM_168

MOM_169

MOM_170

MOM_171

MOM_172

MOM_173

MOM_174

MOM_175

MOM_176

MOM_177

MOM_178

MOM_179

MOM_180

MOM_181

MOM_182

MOM_183

MOM_184

MOM_185

MOM_186

MOM_187

MOM_188

MOM_189

MOM_190

MOM_191

MOM_192

MOM_193

MOM_194

MOM_195

MOM_196

MOM_197

MOM_198

MOM_199

MOM_200

MOM_201

MOM_202

MOM_203

MOM_204

MOM_205

MOM_206

MOM_207

MOM_208

MOM_209

MOM_210

MOM_211

MOM_212

MOM_213

MOM_214

MOM_215

MOM_216
}
\description{
North Atlantic Swordfish Operating Models
 
`MOM_1` is the base case assessment. The rest of the OMs are from the 
OM uncertainty grid.

All OMs are class `MOM` and contain 2 stocks (female and male) and 1 aggregated fleet
}
\keyword{datasets}
