% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{SWO_Data}
\alias{SWO_Data}
\alias{SWOData}
\title{Operating models}
\format{
An object of class \code{Data} of length 1.
}
\usage{
SWOData
}
\description{
Swordfish Data Object
}
\keyword{datasets}
