% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{SWO_MSE}
\alias{SWO_MSE}
\title{Example Swordish MSE}
\format{
An object of class \code{MSE} of length 1.
}
\usage{
SWO_MSE
}
\description{
Example Swordish MSE Object
}
\keyword{datasets}
