% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SWO_userguide}
\alias{SWO_userguide}
\title{Swordfish MSE package user manual}
\usage{
SWO_userguide()
}
\value{
Opens the user guide in browser
}
\description{
Swordfish MSE package user manual
}
\examples{
\dontrun{
SWO_userguide()
}
}
