% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWOMSE.R
\docType{package}
\name{SWOMSE}
\alias{SWOMSE}
\title{SWOMSE: A package for conducting management strategy evaluation (MSE) for
North Atlantic Swordfish}
\description{
Type `SWO_userguide()` for detailed instructions on using the package.
}
\author{
A. Hordyk
}
