% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{Trade_Off}
\alias{Trade_Off}
\title{Create a Trade-Off Plot}
\usage{
Trade_Off(
  DF,
  PM1,
  PM2,
  xline = NULL,
  yline = NULL,
  inc.leg = TRUE,
  pt.size = 2,
  xmax = NULL,
  xmin = NULL,
  ymax = NULL,
  ymin = NULL,
  lab.MPs = NULL
)
}
\arguments{
\item{DF}{A data.frame from `get_tune_OM`}

\item{inc.leg}{logical. Include the legend?}

\item{PMs}{Character length 2. Names of `PM` functions to plot}
}
\value{
A `ggplot` object
}
\description{
Create a Trade-Off Plot
}
