% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{TradeOff}
\alias{TradeOff}
\title{Create a Trade-Off Plot}
\usage{
TradeOff(PM_vals, PMs, xlim = NULL, ylim = NULL, vline = NULL, hline = NULL)
}
\arguments{
\item{PM_vals}{A data.frame generated by `PM_table`}

\item{PMs}{Character length 2. Names of `PM` functions to plot}

\item{xlim}{Optional. Numeric length 1. Minimum value for x}

\item{ylim}{Optional. Numeric length 1. Minimum value for y}
}
\value{
A `ggplot` object
}
\description{
Create a Trade-Off Plot
}
