% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Trim_Data}
\alias{Trim_Data}
\title{Trim Data object}
\usage{
Trim_Data(Data, year)
}
\arguments{
\item{Data}{An object of class `Data`}

\item{year}{The year to trim the data to}
}
\value{
An object of class `Data`
}
\description{
Trim Data object
}
