% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{TS_plot}
\alias{TS_plot}
\title{Time-series plot of SB/SBMSY and TAC}
\usage{
TS_plot(MMSE, relY = FALSE, incFMSY = FALSE)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{relY}{Logical. Make TAC relative to MSY?}
}
\value{
A `ggplot` object
}
\description{
Time-series plot of SB/SBMSY and TAC
}
