% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{TSBio_List}
\alias{TSBio_List}
\title{Time-Series Data of Predicted Spawning Biomass and Recruitment}
\format{
An object of class \code{list} of length 288.
}
\usage{
TSBio_List
}
\description{
A list of predicted spawning biomass and recruitment for each OM
}
\keyword{datasets}
