% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{TSplots}
\alias{TSplots}
\alias{F_FMSY_TS}
\alias{SB_SBMSY_TS}
\alias{Kobe}
\alias{TAC_TS}
\alias{Catch_TS}
\alias{Var_TS}
\title{Plot time-series of Projections}
\usage{
F_FMSY_TS(MMSE, year_range = NULL, mp = NA, fill = "GK", ptsize = 2, ncol = 3)

SB_SBMSY_TS(
  MMSE,
  year_range = NULL,
  mp = NA,
  fill = "GK",
  ref = 1,
  ptsize = 2,
  ncol = 3
)

Kobe(
  MMSE,
  year_range = NULL,
  mp = NA,
  fill = "GK",
  ptsize = 2,
  ncol = 3,
  alpha = 0.1
)

TAC_TS(MMSE, year_range = NULL, mp = NA, hline = NULL)

Catch_TS(MMSE, year_range = NULL, mp = NA, hline = NULL)

Var_TS(MMSE, mp = NA)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{year_range}{Numeric calendar years}

\item{mp}{Optional. Names of the MPs to include}

\item{fill}{Green Kobe 'GK', Limit Reference Point 'LRP', of Overfishing 'OF'}

\item{ptsize}{Size of points}

\item{hline}{Optional. Horizontal line to indicate summary statistic. Numeric length MP.}
}
\value{
A `ggplot` object
}
\description{
Plot time-series of Projections
}
\section{Functions}{
\itemize{
\item \code{F_FMSY_TS()}: Plot time-series of F/FMSY

\item \code{SB_SBMSY_TS()}: Plot time-series of SB/SBMSY

\item \code{Kobe()}: Kobe plot

\item \code{TAC_TS()}: Time-series of projected TAC

\item \code{Catch_TS()}: Time-series of projected catch

\item \code{Var_TS()}: Time-series of percent change in TAC between management cycles

}}
