% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Tune}
\alias{Tune}
\title{Tune an MP}
\usage{
Tune(
  MP_name,
  Tuning_OMs,
  TuneTarget,
  Hist_dir = "Hist_Objects",
  Tune_dir = "Tuning_Objects",
  test_vals = NULL,
  parallel = TRUE,
  maxit = 10,
  step = 0.8,
  tol = 0.001,
  rnd = 3
)
}
\arguments{
\item{MP_name}{Name of an `MP` function}

\item{Tuning_OMs}{Names of the `MOM` objects to tune over}

\item{TuneTarget}{A data.frame of tuning metric and target}

\item{Hist_dir}{Directory where `.hist` objects are saved}

\item{Tune_dir}{Directory where scoping results will be saved}

\item{test_vals}{Optional test values to use}

\item{parallel}{Logical. Run in parallel? (recommended)}

\item{maxit}{maximum number of iterations}

\item{step}{step size}

\item{tol}{tolerance level}

\item{rnd}{rounding value}
}
\value{
saves results to `Tune_dir` and insivibly returns data.frame
}
\description{
Tune an MP
}
