% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{TuneCMP}
\alias{TuneCMP}
\title{Tune a Candidate Management Procedure with 1 Tuning Parameter}
\usage{
TuneCMP(
  MOM_objects = NULL,
  histdir = NULL,
  MPname = NULL,
  tuneMetric = NULL,
  tunetarg = 1,
  step = 0.8,
  maxit = 10,
  tol = 0.001,
  rnd = 3,
  silent = TRUE,
  parallel = TRUE,
  savedir = NULL,
  start_vals = NULL,
  test_vals = NULL
)
}
\arguments{
\item{MOM_objects}{Character vector. Names of the `MOM` objects to tune over}

\item{histdir}{The file path to the directory where the multiHist objects are saved. The files in `histdir` must be `multiHist` objects with
names corresponding to the `MOM_objects` and file extension `.mhist`}

\item{MPname}{Character. Name of the CMP to tune}

\item{tuneMetric}{Name of a PM function}

\item{tunetarg}{The target value for `tuneMetric` output}

\item{step}{Step size for new proposed tuning values}

\item{maxit}{Maximum number of iterations}

\item{tol}{Stopping tolerance}

\item{rnd}{Numeric value for rounding displayed numbers}

\item{silent}{Logical. Hide the output while running closed-loop simulations?}

\item{parallel}{Run in parallel?}

\item{savedir}{Optional. Directory path to save the MP tuning object}

\item{test_vals}{Optional. Numeric vector length 6. Increasing values. Wider range initial tuning parameters.}

\item{start_vals.}{Optional. Numeric vector length 3. Increasing values. Initial tuning parameters.}
}
\description{
Tune a Candidate Management Procedure with 1 Tuning Parameter
}
