% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{TuneMP_1_Par}
\alias{TuneMP_1_Par}
\title{Tune a Candidate Management Procedure with 1 Tuning Parameter}
\usage{
TuneMP_1_Par(
  multiHist = NULL,
  init_vals = NULL,
  MP = NULL,
  tunefunc = NULL,
  tunetarg = 1,
  step = 0.8,
  maxit = 10,
  tol = 0.001,
  rnd = 3,
  silent = TRUE
)
}
\arguments{
\item{multiHist}{An object of class `multiList` or a list of `multiList` objects}

\item{init_vals}{Initial values for the tuning parameter. Numeric values length 3}

\item{MP}{A function of class `MP`}

\item{tunefunc}{A function that takes an object of class `MMSE` and returns a numeric value}

\item{tunetarg}{The target value for `tunefunc` output}

\item{step}{Step size for new proposed tuning values}

\item{maxit}{Maximum number of iterations}

\item{tol}{Stopping tolerance}

\item{rnd}{Numeric value for rounding displayed numbers}

\item{silent}{Logical. Hide the output while running closed-loop simulations?}
}
\description{
Tune a Candidate Management Procedure with 1 Tuning Parameter
}
