% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMSE_plots.R
\name{VarC_Violin}
\alias{VarC_Violin}
\title{Violin Plot for Median Absolute Change in TAC}
\usage{
VarC_Violin(MMSE, np = 1)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}
}
\value{
A `ggplot` object
}
\description{
Violin Plot for Median Absolute Change in TAC
}
