% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_AH.R
\name{WA1}
\alias{WA1}
\alias{WA1_a}
\title{WA 1}
\usage{
WA1(
  x,
  Data,
  Index_ID = c(2, 3, 4, 6),
  Index_CV = c(0.322, 0.342, 0.258, 0.205),
  Data_Lag = 2,
  Interval = 3,
  yrsmth = 3,
  mc = 0.5,
  tunepar = 1,
  ...
)

WA1_a(
  x,
  Data,
  Index_ID = c(2, 3, 4, 6),
  Index_CV = c(0.322, 0.342, 0.258, 0.205),
  Data_Lag = 2,
  Interval = 3,
  yrsmth = 3,
  mc = 0.5,
  tunepar = 1.2,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{tunepar}{Parameter used for tuning}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
WA 1
}
\section{Functions}{
\itemize{
\item \code{WA1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

}}
