% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{BuildData}
\alias{BuildData}
\title{Add the SWO Data Excel file and OM Doc file to the package}
\usage{
BuildData(redoOM = FALSE, redoMSE = FALSE)
}
\arguments{
\item{redoMSE}{Logical. Re-build MSE object?}

\item{redom}{Logical. Re-build `plot(SWOM)`?}
}
\value{
nothing
}
\description{
Add the SWO Data Excel file and OM Doc file to the package
}
\keyword{internal}
