% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{calc_Performance}
\alias{calc_Performance}
\title{Get Performance Metrics from MMSE Object}
\usage{
calc_Performance(MMSE, Class = NULL, PMs = NULL, msg = TRUE)
}
\arguments{
\item{MMSE}{}
}
\description{
Get Performance Metrics from MMSE Object
}
