% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{CPUE_List}
\alias{CPUE_List}
\title{CPUE Fitting Information}
\format{
An object of class \code{list} of length 288.
}
\usage{
CPUE_List
}
\description{
A list of fitted CPUE indices for each OM
}
\keyword{datasets}
