% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Document_MP}
\alias{Document_MP}
\title{Document an MP}
\usage{
Document_MP(MP_name, TuneTargets, Tune_dir = "Tuning_Objects")
}
\arguments{
\item{MP_name}{Name of an `MP` function}

\item{TuneTargets}{A data.frame of tuning metrics and targets}

\item{Tune_dir}{Directory where scoping results are saved}
}
\value{
Nothing
}
\description{
Document an MP
}
