% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen_OMs.r
\docType{data}
\name{Fleet_DF}
\alias{Fleet_DF}
\title{Fleet Information}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\usage{
Fleet_DF
}
\description{
A dataframe of information on fleet names
}
\keyword{datasets}
