% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_AH.R
\name{IR1}
\alias{IR1}
\alias{CI1_a}
\alias{CA1_a}
\alias{CT1_a}
\alias{JP1_a}
\alias{US1_a}
\alias{IR1_a}
\alias{PO1_a}
\alias{MO1_a}
\title{Index Ratio 1}
\usage{
IR1(
  x,
  Data,
  Index_ID = 1,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.9,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

CI1_a(
  x,
  Data,
  Index_ID = 1,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.849,
  mc = 0.5,
  useCombined = TRUE,
  yrsmth = 3,
  ...
)

CA1_a(
  x,
  Data,
  Index_ID = 2,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.974,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

CT1_a(
  x,
  Data,
  Index_ID = 4,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.873,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

JP1_a(
  x,
  Data,
  Index_ID = 3,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.858,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

US1_a(
  x,
  Data,
  Index_ID = 6,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.853,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

IR1_a(
  x,
  Data,
  Index_ID = 1,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 1,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

PO1_a(
  x,
  Data,
  Index_ID = 7,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.815,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)

MO1_a(
  x,
  Data,
  Index_ID = 5,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 0.749,
  mc = 0.5,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
This MP adjusts the TAC based on ratio of the mean index over the last 3 years (y-2):y
to the mean index over the 3 years before that (y-5):(y-3)
}
\section{Functions}{
\itemize{
\item \code{CI1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{CA1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{CT1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{JP1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{US1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{IR1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{PO1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

\item \code{MO1_a()}: Tuned to PGK_6_10 = 0.60 across Reference OMs.

}}
