% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuning.R
\name{Plot_Tune}
\alias{Plot_Tune}
\title{Plot the results of tuning for an MP}
\usage{
Plot_Tune(MP_name, Tune_dir = "Tuning_Objects", plot = TRUE)
}
\arguments{
\item{MP_name}{Name of the MP}

\item{Tune_dir}{Directory where scoping results are saved}

\item{plot}{Logical. Show the plot?}
}
\value{
list ggplot object and data.frame
}
\description{
Plot the results of tuning for an MP
}
