% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{PMs}
\alias{PMs}
\alias{PGK_10}
\alias{PGK_med}
\alias{PGK_long}
\alias{PGK}
\alias{PGK_30}
\alias{POF}
\alias{PNOF}
\alias{LRP_short}
\alias{LRP_long}
\alias{LRP}
\alias{nLRP_short}
\alias{nLRP_long}
\alias{nLRP}
\alias{TAC1}
\alias{AvC10}
\alias{AvTAC10}
\alias{AvC30}
\alias{AvTAC30}
\alias{VarC}
\alias{MaxVarC}
\title{Swordfish Performance Metrics functions}
\usage{
PGK_10(MMSEobj = NULL, Ref = 1, Yrs = c(13, 13))

PGK_med(MMSEobj = NULL, Ref = 1, Yrs = c(9, 13))

PGK_long(MMSEobj = NULL, Ref = 1, Yrs = c(14, 33))

PGK(MMSEobj = NULL, Ref = 1, Yrs = c(4, 33))

PGK_30(MMSEobj = NULL, Ref = 1, Yrs = c(33, 33))

POF(MMSEobj = NULL, Ref = 1, Yrs = c(4, 33))

PNOF(MMSEobj = NULL, Ref = 1, Yrs = c(4, 33))

LRP_short(MMSEobj = NULL, Ref = 0.4, Yrs = c(4, 13))

LRP_long(MMSEobj = NULL, Ref = 0.4, Yrs = c(14, 33))

LRP(MMSEobj = NULL, Ref = 0.4, Yrs = c(4, 33))

nLRP_short(MMSEobj = NULL, Ref = 0.4, Yrs = c(4, 13))

nLRP_long(MMSEobj = NULL, Ref = 0.4, Yrs = c(14, 33))

nLRP(MMSEobj = NULL, Ref = 0.4, Yrs = c(4, 33))

TAC1(MMSEobj = NULL, Ref = 1, Yrs = c(4, 4))

AvC10(MMSEobj = NULL, Ref = NULL, Yrs = c(4, 13))

AvTAC10(MMSEobj = NULL, Ref = NULL, Yrs = c(4, 13))

AvC30(MMSEobj = NULL, Ref = 1, Yrs = c(14, 33))

AvTAC30(MMSEobj = NULL, Ref = 1, Yrs = c(14, 33))

VarC(MMSEobj = NULL, Ref = 1, Yrs = c(4, 33))

MaxVarC(MMSEobj = NULL, Ref = 1, Yrs = c(4, 33))
}
\arguments{
\item{MMSEobj}{An object of class `MMSE`}

\item{Ref}{Reference point used in the performance metrics (e.g., 0.5BMSY)}

\item{Yrs}{Years the performance metric is calculated over}
}
\value{
An object of class `PM`
}
\description{
Swordfish Performance Metrics functions
}
\section{Functions}{
\itemize{
\item \code{PGK_10()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Year 10 (2033)

\item \code{PGK_med()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Years 6-10 (2029-2033)

\item \code{PGK_long()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Years 11-30 (2034-2053)

\item \code{PGK()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) over all years (2024-2053)

\item \code{PGK_30()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) in Year 30 (2053)

\item \code{POF()}: Probability of Overfishing (F>FMSY) over all years (2024-2053)

\item \code{PNOF()}: Probability of Not Overfishing (F<FMSY) over all years (2024-2053)

\item \code{LRP_short()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any of the first 10 years (2024-2033)

\item \code{LRP_long()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any of years 11-30 (2034-2053)

\item \code{LRP()}: Probability of breaching the limit reference point (SSB<0.4SSB_MSY) in any year (2024-2053)

\item \code{nLRP_short()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any of the first 10 years (2024-2033)

\item \code{nLRP_long()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any of years 11-30 (2034-2053)

\item \code{nLRP()}: Probability of not breaching the limit reference point (SSB>0.4SSB_MSY) in any year (2024-2053)

\item \code{TAC1()}: TAC in the First Implementation Year (2024)

\item \code{AvC10()}: Median catches (t) over years 1-10 (2024-2033)

\item \code{AvTAC10()}: Median TAC (t) over years 1-10 (2024-2033)

\item \code{AvC30()}: Median catches (t) over years 11-30 (2034-2053)

\item \code{AvTAC30()}: Median TAC (t) over years 11-30 (2034-2053)

\item \code{VarC()}: Median variation in TAC (\%) between management cycles over all years

\item \code{MaxVarC()}: Maximum variation in TAC (\%) between management cycles over all years

}}
\concept{Safety}
\concept{Stability}
\concept{Status}
\concept{Yield}
