% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{PMs}
\alias{PMs}
\alias{Safety_S}
\alias{Safety_M}
\alias{Status_S}
\alias{Status_M}
\alias{Stability}
\alias{Yield_S}
\alias{Yield_M}
\title{Swordfish Performance Metrics functions}
\usage{
Safety_S(MSEobj = NULL, Ref = 0.4, Yrs = 10)

Safety_M(MSEobj = NULL, Ref = 0.4, Yrs = c(11, 30))

Status_S(MSEobj = NULL, Ref = 1, Yrs = 10)

Status_M(MSEobj = NULL, Ref = 1, Yrs = c(11, 30))

Stability(MSEobj = NULL, Ref = 1, Yrs = NULL)

Yield_S(MSEobj = NULL, Ref = 1, Yrs = 10)

Yield_M(MSEobj = NULL, Ref = 1, Yrs = c(11, 30))
}
\arguments{
\item{MSEobj}{An object of class `MSE`}

\item{Ref}{Reference point used in the performance metrics (e.g., 0.5BMSY)}

\item{Yrs}{Years the performance metric is calculated over}
}
\value{
An object of class `PM`
}
\description{
Swordfish Performance Metrics functions
}
\section{Functions}{
\itemize{
\item \code{Safety_S}: Probability Spawning Biomass is > 0.4SBMSY over first 10 years

\item \code{Safety_M}: Probability Spawning Biomass is > 0.4SBMSY over last 20 years

\item \code{Status_S}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) over first  10 years

\item \code{Status_M}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) over last  20 years

\item \code{Stability}: Average Annual Variability in Yield

\item \code{Yield_S}: Average Yield in first 10 years

\item \code{Yield_M}: Average Yield in last 20 years
}}

