% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{PMs}
\alias{PMs}
\alias{Safety_S}
\alias{Safety_M}
\alias{Status_S}
\alias{Status_M}
\alias{Stability}
\alias{Yield_S}
\alias{Yield_M}
\title{Swordfish Performance Metrics functions}
\usage{
Safety_S(MMSEobj = NULL, Ref = 0.4, Yrs = 10)

Safety_M(MMSEobj = NULL, Ref = 0.4, Yrs = c(11, 30))

Status_S(MMSEobj = NULL, Ref = 1, Yrs = 10)

Status_M(MMSEobj = NULL, Ref = 1, Yrs = c(11, 30))

Stability(MMSEobj = NULL, Ref = 0.25, Yrs = NULL)

Yield_S(MMSEobj = NULL, Ref = 1, Yrs = 10)

Yield_M(MMSEobj = NULL, Ref = 1, Yrs = c(11, 30))
}
\arguments{
\item{MMSEobj}{An object of class `MMSE`}

\item{Ref}{Reference point used in the performance metrics (e.g., 0.5BMSY)}

\item{Yrs}{Years the performance metric is calculated over}
}
\value{
An object of class `PM`
}
\description{
Swordfish Performance Metrics functions
}
\section{Functions}{
\itemize{
\item \code{Safety_S()}: Probability Female Spawning Biomass is > 0.4SBMSY over first 10 years

\item \code{Safety_M()}: Probability Female Spawning Biomass is > 0.4SBMSY over last 20 years

\item \code{Status_S()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) over first  10 years

\item \code{Status_M()}: Probability of being in Green Zone of Kobe Space (SB>SBMSY & F<FMSY) over last  20 years

\item \code{Stability()}: Probability Annual Variability in Yield < 25%

\item \code{Yield_S()}: Average Yield in first 10 years

\item \code{Yield_M()}: Average Yield in last 20 years

}}
