% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs.R
\name{SameTAC}
\alias{SameTAC}
\title{Should the TAC remain unchanged?}
\usage{
SameTAC(Initial_MP_Yr, Interval, Data)
}
\arguments{
\item{Initial_MP_Yr}{The calendar year when the MP will first be implemented}

\item{Interval}{The management interval where the TAC is updated}

\item{Data}{An object of class `Data`}
}
\value{
TRUE if the TAC should remain unchanged and FALSE otherwise
}
\description{
Should the TAC remain unchanged?
}
