% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP1.R, R/SP2.R, R/SP2un.R, R/SP3.R
\name{SP1}
\alias{SP1}
\alias{SP1_a}
\alias{SP1_b}
\alias{SP1_c}
\alias{SP1_e}
\alias{SP1_g}
\alias{SP1_h}
\alias{SP2}
\alias{SP2un}
\alias{SP3}
\title{A Surplus Production model with a Harvest Control Rule that linearly reduces F}
\usage{
SP1(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)

SP1_a(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 3.11318681318681,
  mc = 0.25,
  ...
)

SP1_b(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 3.38664343786295,
  mc = 0.25,
  ...
)

SP1_c(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 3.61603053435114,
  mc = 0.25,
  ...
)

SP1_e(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 3.33004926108374,
  mc = 0.25,
  ...
)

SP1_g(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 3.87857142857143,
  mc = 0.25,
  ...
)

SP1_h(
  x,
  Data,
  Data_Lag = 1,
  Interval = 3,
  tunepar = 3.32857142857143,
  mc = 0.25,
  ...
)

SP2(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)

SP2un(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = NA, ...)

SP3(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
The SP model assumes a Schaefer production model
}
\section{Functions}{
\itemize{
\item \code{SP1_a()}: Tuned to PGK_6_10 = 0.6 across Reference OMs.

\item \code{SP1_b()}: Tuned to PGK_med = 0.6 across Reference OMs.

\item \code{SP1_c()}: Tuned to PGK_long = 0.6 across Reference OMs.

\item \code{SP1_e()}: Tuned to PGK_6_10 = 0.51 across Reference OMs.

\item \code{SP1_g()}: Tuned to LRP = 0.1 across Reference OMs.

\item \code{SP1_h()}: Tuned to LRP = 0.05 across Reference OMs.

\item \code{SP2()}: Same as SP1, but uses the Fox production model

\item \code{SP2un()}: Same as SP1, but uses the Fox production model

\item \code{SP3()}: Fox production model with

}}
