% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP1.R, R/SP2.R
\name{SP1}
\alias{SP1}
\alias{SP1_a}
\alias{SP1_b}
\alias{SP1_c}
\alias{SP2}
\title{A Surplus Production model with a Harvest Control Rule that linearly reduces F}
\usage{
SP1(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)

SP1_a(x, Data, Data_Lag = 1, Interval = 3, tunepar = 2.404, mc = 0.25, ...)

SP1_b(x, Data, Data_Lag = 1, Interval = 3, tunepar = 3.174, mc = 0.25, ...)

SP1_c(x, Data, Data_Lag = 1, Interval = 3, tunepar = 3.518, mc = 0.25, ...)

SP2(x, Data, Data_Lag = 1, Interval = 3, tunepar = 1, mc = 0.25, ...)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
The SP model assumes a Schaefer production model
}
\section{Functions}{
\itemize{
\item \code{SP1_a()}: Tuned to PGK_med = 0.6 across Reference OMs.

\item \code{SP1_b()}: Tuned to PGK_long = 0.6 across Reference OMs.

\item \code{SP1_c()}: Tuned to PGK_30 = 0.6 across Reference OMs.

\item \code{SP2()}: Same as SP1, but uses the Fox production model

}}
