% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPs_AH.R
\name{SP1}
\alias{SP1}
\title{Index Ratio 1}
\usage{
SP1(
  x,
  Data,
  Index_ID = 1,
  Data_Lag = 2,
  Interval = 3,
  tunepar = 1,
  mc = 0.25,
  useCombined = FALSE,
  yrsmth = 3,
  ...
)
}
\arguments{
\item{x}{A position in the data object}

\item{Data}{An object of class `Data`}

\item{Data_Lag}{The number of years to lag the data}

\item{Interval}{The TAC update interval}

\item{tunepar}{Parameter used for tuning}

\item{mc}{The maximum fractional change in the TAC among years. NA to ignore}

\item{...}{Additional arguments (unused)}

\item{yrs}{numeric vector length 2. The years to calculate the mean index. Ratio is
calculated as mean(ind[yrs[1]])/mean(ind[yrs[2]])}
}
\value{
An object of class `Rec` with the `TAC` slot populated
}
\description{
This MP adjusts the TAC based on ratio of the mean index over the last 3 years (y-2):y
to the mean index over the 3 years before that (y-5):(y-3)
}
