% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen.r
\docType{data}
\name{SWO-OMs}
\alias{SWO-OMs}
\alias{MOM_000}
\alias{MOM_133}
\alias{MOM_134}
\alias{MOM_135}
\alias{MOM_139}
\alias{MOM_140}
\alias{MOM_141}
\alias{MOM_127}
\alias{MOM_128}
\alias{MOM_129}
\alias{MOM_136}
\alias{MOM_137}
\alias{MOM_138}
\alias{MOM_142}
\alias{MOM_143}
\alias{MOM_144}
\alias{MOM_151}
\alias{MOM_152}
\alias{MOM_153}
\alias{MOM_157}
\alias{MOM_158}
\alias{MOM_159}
\alias{MOM_187}
\alias{MOM_188}
\alias{MOM_189}
\alias{MOM_193}
\alias{MOM_194}
\alias{MOM_195}
\title{Operating models}
\format{
An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.
}
\usage{
MOM_000

MOM_133

MOM_134

MOM_135

MOM_139

MOM_140

MOM_141

MOM_127

MOM_128

MOM_129

MOM_136

MOM_137

MOM_138

MOM_142

MOM_143

MOM_144

MOM_151

MOM_152

MOM_153

MOM_157

MOM_158

MOM_159

MOM_187

MOM_188

MOM_189

MOM_193

MOM_194

MOM_195
}
\description{
North Atlantic Swordfish Operating Models
 
`MOM_000` is the base case assessment. The rest of the OMs are from the 
OM uncertainty grid.

All OMs are class `MOM` and contain 2 stocks (female and male) and 1 aggregated fleet
}
\keyword{datasets}
