% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen.r
\docType{data}
\name{SWO-OMs}
\alias{SWO-OMs}
\alias{MOM_001}
\alias{MOM_002}
\alias{MOM_003}
\alias{MOM_004}
\alias{MOM_005}
\alias{MOM_006}
\alias{MOM_007}
\alias{MOM_008}
\alias{MOM_009}
\alias{MOM_010}
\alias{MOM_011}
\title{Operating models}
\format{
An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.

An object of class \code{MOM} of length 1.
}
\usage{
MOM_001

MOM_002

MOM_003

MOM_004

MOM_005

MOM_006

MOM_007

MOM_008

MOM_009

MOM_010

MOM_011
}
\description{
North Atlantic Swordfish Operating Models
 
`MOM_000` is the base case assessment. The rest of the OMs are from the 
OM uncertainty grid.

All OMs are class `MOM` and contain 2 stocks (female and male) and 1 aggregated fleet
}
\keyword{datasets}
