% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxygen.r
\docType{data}
\name{SWO-OMs}
\alias{SWO-OMs}
\alias{OM_1}
\alias{OM_2}
\alias{OM_3}
\alias{OM_4}
\alias{OM_5}
\alias{OM_6}
\alias{OM_7}
\alias{OM_8}
\alias{OM_9}
\alias{OM_10}
\alias{OM_11}
\alias{OM_12}
\alias{OM_13}
\alias{OM_14}
\alias{OM_15}
\alias{OM_16}
\alias{OM_17}
\alias{OM_18}
\alias{OM_19}
\alias{OM_20}
\alias{OM_21}
\alias{OM_22}
\alias{OM_23}
\alias{OM_24}
\alias{OM_25}
\alias{OM_26}
\alias{OM_27}
\alias{OM_28}
\alias{OM_29}
\alias{OM_30}
\alias{OM_31}
\alias{OM_32}
\alias{OM_33}
\alias{OM_34}
\alias{OM_35}
\alias{OM_36}
\alias{OM_37}
\alias{OM_38}
\alias{OM_39}
\alias{OM_40}
\alias{OM_41}
\alias{OM_42}
\alias{OM_43}
\alias{OM_44}
\alias{OM_45}
\alias{OM_46}
\alias{OM_47}
\alias{OM_48}
\alias{OM_49}
\alias{OM_50}
\alias{OM_51}
\alias{OM_52}
\alias{OM_53}
\alias{OM_54}
\alias{OM_55}
\alias{OM_56}
\alias{OM_57}
\alias{OM_58}
\alias{OM_59}
\alias{OM_60}
\alias{OM_61}
\alias{OM_62}
\alias{OM_63}
\alias{OM_64}
\alias{OM_65}
\alias{OM_66}
\alias{OM_67}
\alias{OM_68}
\alias{OM_69}
\alias{OM_70}
\alias{OM_71}
\alias{OM_72}
\alias{OM_73}
\alias{OM_74}
\alias{OM_75}
\alias{OM_76}
\alias{OM_77}
\alias{OM_78}
\alias{OM_79}
\alias{OM_80}
\alias{OM_81}
\alias{OM_82}
\alias{OM_83}
\alias{OM_84}
\alias{OM_85}
\alias{OM_86}
\alias{OM_87}
\alias{OM_88}
\alias{OM_89}
\alias{OM_90}
\alias{OM_91}
\alias{OM_92}
\alias{OM_93}
\alias{OM_94}
\alias{OM_95}
\alias{OM_96}
\alias{OM_97}
\alias{OM_98}
\alias{OM_99}
\alias{OM_100}
\alias{OM_101}
\alias{OM_102}
\alias{OM_103}
\alias{OM_104}
\alias{OM_105}
\alias{OM_106}
\alias{OM_107}
\alias{OM_108}
\alias{OM_109}
\alias{OM_110}
\alias{OM_111}
\alias{OM_112}
\alias{OM_113}
\alias{OM_114}
\alias{OM_115}
\alias{OM_116}
\alias{OM_117}
\alias{OM_118}
\alias{OM_119}
\alias{OM_120}
\alias{OM_121}
\alias{OM_122}
\alias{OM_123}
\alias{OM_124}
\alias{OM_125}
\alias{OM_126}
\alias{OM_127}
\alias{OM_128}
\alias{OM_129}
\alias{OM_130}
\alias{OM_131}
\alias{OM_132}
\alias{OM_133}
\alias{OM_134}
\alias{OM_135}
\alias{OM_136}
\alias{OM_137}
\alias{OM_138}
\alias{OM_139}
\alias{OM_140}
\alias{OM_141}
\alias{OM_142}
\alias{OM_143}
\alias{OM_144}
\alias{OM_145}
\alias{OM_146}
\alias{OM_147}
\alias{OM_148}
\alias{OM_149}
\alias{OM_150}
\alias{OM_151}
\alias{OM_152}
\alias{OM_153}
\alias{OM_154}
\alias{OM_155}
\alias{OM_156}
\alias{OM_157}
\alias{OM_158}
\alias{OM_159}
\alias{OM_160}
\alias{OM_161}
\alias{OM_162}
\alias{OM_163}
\alias{OM_164}
\alias{OM_165}
\alias{OM_166}
\alias{OM_167}
\alias{OM_168}
\alias{OM_169}
\alias{OM_170}
\alias{OM_171}
\alias{OM_172}
\alias{OM_173}
\alias{OM_174}
\alias{OM_175}
\alias{OM_176}
\alias{OM_177}
\alias{OM_178}
\alias{OM_179}
\alias{OM_180}
\alias{OM_181}
\alias{OM_182}
\alias{OM_183}
\alias{OM_184}
\alias{OM_185}
\alias{OM_186}
\alias{OM_187}
\alias{OM_188}
\alias{OM_189}
\alias{OM_190}
\alias{OM_191}
\alias{OM_192}
\alias{OM_193}
\alias{OM_194}
\alias{OM_195}
\alias{OM_196}
\alias{OM_197}
\alias{OM_198}
\alias{OM_199}
\alias{OM_200}
\alias{OM_201}
\alias{OM_202}
\alias{OM_203}
\alias{OM_204}
\alias{OM_205}
\alias{OM_206}
\alias{OM_207}
\alias{OM_208}
\alias{OM_209}
\alias{OM_210}
\alias{OM_211}
\alias{OM_212}
\alias{OM_213}
\alias{OM_214}
\alias{OM_215}
\alias{OM_216}
\title{Operating models}
\format{
An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.

An object of class \code{OM} of length 1.
}
\usage{
OM_1

OM_2

OM_3

OM_4

OM_5

OM_6

OM_7

OM_8

OM_9

OM_10

OM_11

OM_12

OM_13

OM_14

OM_15

OM_16

OM_17

OM_18

OM_19

OM_20

OM_21

OM_22

OM_23

OM_24

OM_25

OM_26

OM_27

OM_28

OM_29

OM_30

OM_31

OM_32

OM_33

OM_34

OM_35

OM_36

OM_37

OM_38

OM_39

OM_40

OM_41

OM_42

OM_43

OM_44

OM_45

OM_46

OM_47

OM_48

OM_49

OM_50

OM_51

OM_52

OM_53

OM_54

OM_55

OM_56

OM_57

OM_58

OM_59

OM_60

OM_61

OM_62

OM_63

OM_64

OM_65

OM_66

OM_67

OM_68

OM_69

OM_70

OM_71

OM_72

OM_73

OM_74

OM_75

OM_76

OM_77

OM_78

OM_79

OM_80

OM_81

OM_82

OM_83

OM_84

OM_85

OM_86

OM_87

OM_88

OM_89

OM_90

OM_91

OM_92

OM_93

OM_94

OM_95

OM_96

OM_97

OM_98

OM_99

OM_100

OM_101

OM_102

OM_103

OM_104

OM_105

OM_106

OM_107

OM_108

OM_109

OM_110

OM_111

OM_112

OM_113

OM_114

OM_115

OM_116

OM_117

OM_118

OM_119

OM_120

OM_121

OM_122

OM_123

OM_124

OM_125

OM_126

OM_127

OM_128

OM_129

OM_130

OM_131

OM_132

OM_133

OM_134

OM_135

OM_136

OM_137

OM_138

OM_139

OM_140

OM_141

OM_142

OM_143

OM_144

OM_145

OM_146

OM_147

OM_148

OM_149

OM_150

OM_151

OM_152

OM_153

OM_154

OM_155

OM_156

OM_157

OM_158

OM_159

OM_160

OM_161

OM_162

OM_163

OM_164

OM_165

OM_166

OM_167

OM_168

OM_169

OM_170

OM_171

OM_172

OM_173

OM_174

OM_175

OM_176

OM_177

OM_178

OM_179

OM_180

OM_181

OM_182

OM_183

OM_184

OM_185

OM_186

OM_187

OM_188

OM_189

OM_190

OM_191

OM_192

OM_193

OM_194

OM_195

OM_196

OM_197

OM_198

OM_199

OM_200

OM_201

OM_202

OM_203

OM_204

OM_205

OM_206

OM_207

OM_208

OM_209

OM_210

OM_211

OM_212

OM_213

OM_214

OM_215

OM_216
}
\description{
North Atlantic Swordfish Operating Models
 
`SWOM` is a single OM combining all SS OMs including `OM_base_case` and
all SS OMs there were imported into the package (`OM_1`, `OM_2`, ..., etc)
}
\keyword{datasets}
