% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roxy_DataObjects.r
\docType{data}
\name{SWO-OMs}
\alias{SWO-OMs}
\alias{SWOM}
\alias{OM_base_case}
\alias{OM_19}
\alias{OM_20}
\alias{OM_21}
\alias{OM_22}
\alias{OM_23}
\alias{OM_24}
\alias{OM_25}
\alias{OM_26}
\alias{OM_27}
\alias{OM_28}
\alias{OM_29}
\alias{OM_30}
\alias{OM_31}
\alias{OM_32}
\alias{OM_33}
\alias{OM_34}
\alias{OM_35}
\alias{OM_36}
\alias{OM_39}
\alias{OM_44}
\alias{OM_45}
\alias{OM_50}
\alias{OM_51}
\alias{OM_55}
\alias{OM_56}
\alias{OM_57}
\alias{OM_58}
\alias{OM_59}
\alias{OM_60}
\alias{OM_61}
\alias{OM_62}
\alias{OM_63}
\alias{OM_64}
\alias{OM_65}
\alias{OM_66}
\alias{OM_67}
\alias{OM_68}
\alias{OM_69}
\alias{OM_70}
\alias{OM_71}
\alias{OM_72}
\alias{OM_91}
\alias{OM_92}
\alias{OM_93}
\alias{OM_94}
\alias{OM_95}
\alias{OM_96}
\alias{OM_97}
\alias{OM_98}
\alias{OM_99}
\alias{OM_100}
\alias{OM_101}
\alias{OM_102}
\alias{OM_103}
\alias{OM_104}
\alias{OM_105}
\alias{OM_106}
\alias{OM_107}
\alias{OM_108}
\alias{OM_111}
\alias{OM_116}
\alias{OM_117}
\alias{OM_121}
\alias{OM_122}
\alias{OM_123}
\alias{OM_127}
\alias{OM_128}
\alias{OM_129}
\alias{OM_130}
\alias{OM_131}
\alias{OM_132}
\alias{OM_133}
\alias{OM_134}
\alias{OM_135}
\alias{OM_136}
\alias{OM_137}
\alias{OM_138}
\alias{OM_139}
\alias{OM_140}
\alias{OM_141}
\alias{OM_142}
\alias{OM_143}
\alias{OM_144}
\alias{OM_163}
\alias{OM_164}
\alias{OM_165}
\alias{OM_166}
\alias{OM_167}
\alias{OM_168}
\alias{OM_169}
\alias{OM_170}
\alias{OM_171}
\alias{OM_172}
\alias{OM_173}
\alias{OM_174}
\alias{OM_175}
\alias{OM_176}
\alias{OM_177}
\alias{OM_178}
\alias{OM_179}
\alias{OM_180}
\alias{OM_181}
\alias{OM_182}
\alias{OM_183}
\alias{OM_187}
\alias{OM_188}
\alias{OM_189}
\alias{OM_193}
\alias{OM_194}
\alias{OM_195}
\alias{OM_199}
\alias{OM_200}
\alias{OM_201}
\alias{OM_202}
\alias{OM_203}
\alias{OM_204}
\alias{OM_205}
\alias{OM_206}
\alias{OM_207}
\alias{OM_208}
\alias{OM_209}
\alias{OM_210}
\alias{OM_211}
\alias{OM_212}
\alias{OM_213}
\alias{OM_214}
\alias{OM_215}
\alias{OM_216}
\alias{OM_235}
\alias{OM_236}
\alias{OM_237}
\alias{OM_238}
\alias{OM_239}
\alias{OM_240}
\alias{OM_241}
\alias{OM_242}
\alias{OM_243}
\alias{OM_244}
\alias{OM_245}
\alias{OM_246}
\alias{OM_247}
\alias{OM_248}
\alias{OM_249}
\alias{OM_250}
\alias{OM_251}
\alias{OM_252}
\alias{OM_253}
\alias{OM_254}
\alias{OM_255}
\alias{OM_259}
\alias{OM_260}
\alias{OM_261}
\alias{OM_265}
\alias{OM_266}
\alias{OM_267}
\alias{OM_271}
\alias{OM_272}
\alias{OM_273}
\alias{OM_274}
\alias{OM_275}
\alias{OM_276}
\alias{OM_277}
\alias{OM_278}
\alias{OM_279}
\alias{OM_280}
\alias{OM_281}
\alias{OM_282}
\alias{OM_283}
\alias{OM_284}
\alias{OM_285}
\alias{OM_286}
\alias{OM_287}
\alias{OM_288}
\title{Operating models}
\format{An object of class \code{OM} of length 1.}
\usage{
SWOM

OM_base_case

OM_19

OM_20

OM_21

OM_22

OM_23

OM_24

OM_25

OM_26

OM_27

OM_28

OM_29

OM_30

OM_31

OM_32

OM_33

OM_34

OM_35

OM_36

OM_39

OM_44

OM_45

OM_50

OM_51

OM_55

OM_56

OM_57

OM_58

OM_59

OM_60

OM_61

OM_62

OM_63

OM_64

OM_65

OM_66

OM_67

OM_68

OM_69

OM_70

OM_71

OM_72

OM_91

OM_92

OM_93

OM_94

OM_95

OM_96

OM_97

OM_98

OM_99

OM_100

OM_101

OM_102

OM_103

OM_104

OM_105

OM_106

OM_107

OM_108

OM_111

OM_116

OM_117

OM_121

OM_122

OM_123

OM_127

OM_128

OM_129

OM_130

OM_131

OM_132

OM_133

OM_134

OM_135

OM_136

OM_137

OM_138

OM_139

OM_140

OM_141

OM_142

OM_143

OM_144

OM_163

OM_164

OM_165

OM_166

OM_167

OM_168

OM_169

OM_170

OM_171

OM_172

OM_173

OM_174

OM_175

OM_176

OM_177

OM_178

OM_179

OM_180

OM_181

OM_182

OM_183

OM_187

OM_188

OM_189

OM_193

OM_194

OM_195

OM_199

OM_200

OM_201

OM_202

OM_203

OM_204

OM_205

OM_206

OM_207

OM_208

OM_209

OM_210

OM_211

OM_212

OM_213

OM_214

OM_215

OM_216

OM_235

OM_236

OM_237

OM_238

OM_239

OM_240

OM_241

OM_242

OM_243

OM_244

OM_245

OM_246

OM_247

OM_248

OM_249

OM_250

OM_251

OM_252

OM_253

OM_254

OM_255

OM_259

OM_260

OM_261

OM_265

OM_266

OM_267

OM_271

OM_272

OM_273

OM_274

OM_275

OM_276

OM_277

OM_278

OM_279

OM_280

OM_281

OM_282

OM_283

OM_284

OM_285

OM_286

OM_287

OM_288
}
\description{
Swordfish operating models.
 
`SWOM` is a single OM combining all SS OMs including `OM_base_case` and
all SS OMs there were imported into the package (`OM_19`, `OM_20`, ..., `OM_288`)
}
\keyword{datasets}
