% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMs.R
\name{TradeOff}
\alias{TradeOff}
\title{Create a Trade-Off Plot}
\usage{
TradeOff(
  MMSE,
  PMs,
  xlim = NULL,
  ylim = NULL,
  vline = NULL,
  hline = NULL,
  quants = c(0.1, 0.9),
  inc.leg = TRUE,
  subset = FALSE
)
}
\arguments{
\item{MMSE}{An object of class `MMSE`}

\item{PMs}{Character length 2. Names of `PM` functions to plot}

\item{xlim}{Optional. Numeric length 1. Minimum value for x}

\item{ylim}{Optional. Numeric length 1. Minimum value for y}

\item{vline}{Optional. Numeric vector for vertical lines}

\item{hline}{Optional. Numeric vector for horizontal lines}

\item{quants}{Numeric vector length 2 of quantiles for error bars. Only shown
for TAC PMs. Ignored if NULL}

\item{inc.leg}{logical. Include the legend?}
}
\value{
A `ggplot` object
}
\description{
Create a Trade-Off Plot
}
